worker/requirements.txt
faster-whisper>=1.0.0
ctranslate2>=4.3.0
whisperx>=3.1.0
pyannote.audio>=3.1.0
speechbrain>=0.5.16
transformers>=4.41.0
sentence-transformers>=2.6.0
pgvector>=0.2.5
redis>=5.0.6
rq>=1.16.2
httpx>=0.27.0
tiktoken>=0.7.0
python-dotenv>=1.0.1
sqlalchemy>=2.0.30
psycopg[binary]>=3.1.19
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
